Rename `infoHeader` dir to `homepage`
