+++ client/src/main/scripts/components/homepage/infoHeader.ts
